MASTER | fix for going back states bug after the py3 merge BLOCKUTILS | part of the MASTER fix POSTUTILS.autoSwim | allow for multiple setups on one surface - layering setup RIGSHAPES.ik_end | changed out tipMid resolved. It was calculating just from start to end which didn't work for non straight sections. Now does midpoint on a curve along the rig joints of the module DESIGNERSPRING/DRAGGER | bug fix. Same matrix aim bug. Changed over to local README  | added some links

+++ cgm/core/mrs/lib/block_utils.py
@@ -1089,29 +1089,30 @@ def prerig_simple(self):
     return True
 
 def prerig_delete(self, msgLinks = [], msgLists = [], formHandles = True):
-    try:
-        _str_func = 'prerig_delete'
-        log.debug(cgmGEN.logString_start(_str_func))
-        
-        try:self.moduleTarget.delete()
-        except:log.debug("|{0}| >> No moduleTarget...".format(_str_func))
+    _str_func = 'prerig_delete'
+    log.debug(cgmGEN.logString_start(_str_func))
+    
+    try:self.moduleTarget.delete()
+    except:log.debug("|{0}| >> No moduleTarget...".format(_str_func))
+    if self.getMessage('prerigNull'):
         self.prerigNull.delete()
-        if self.getMessage('noTransformNull'):
-            self.noTransformNull.delete()
-        if formHandles:
-            for mHandle in [self] + self.msgList_get('formHandles'):
-                try:mHandle.jointHelper.delete()
-                except:pass    
-        
-        for l in msgLists:
-            _buffer = self.msgList_get(l,asMeta=False)
-            if not _buffer:
-                log.debug("|{0}| >> Missing msgList: {1}".format(_str_func,l))  
-            else:
-                mc.delete(_buffer)
-        return True   
-    except Exception as err:
-        cgmGEN.cgmExceptCB(Exception,err)
+    else:
+        log.warning("No prerigNull found")
+    if self.getMessage('noTransformNull'):
+        self.noTransformNull.delete()
+    if formHandles:
+        for mHandle in [self] + self.msgList_get('formHandles'):
+            try:mHandle.jointHelper.delete()
+            except:pass    
+    
+    for l in msgLists:
+        _buffer = self.msgList_get(l,asMeta=False)
+        if not _buffer:
+            log.debug("|{0}| >> Missing msgList: {1}".format(_str_func,l))  
+        else:
+            mc.delete(_buffer)
+    return True   
+
         
 def prerig_handlesLock(self, lock=None):
     try:
